SDET-Python-final_project/code/tests/code/requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
click==8.0.1
colorama==0.4.4
Faker==4.18.0
Flask==2.0.1
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.1.2
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
requests==2.25.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.3.20
text-unidecode==1.3
toml==0.10.2
tqdm==4.61.0
urllib3==1.26.5
webdrivermanager==0.10.0
Werkzeug==2.0.1
